Drop non-word meeting features from zip export.
